Have an option to skip os deps installation in bootstrap script (#1939)
* Have an option to skip os deps installation in bootstrap script

* Check even before running the os deps install script

* Add check in install_os_deps script to see if boost is installed and version is sufficient

* Added missing equals check

* Build bf-syslibs and bf-utils also from brig-stable refpoint as these repos have been changing as well recently

* Check if the version.hpp file exists on pre-installed boost environment

+++ install_os_deps.sh
@@ -187,16 +187,33 @@ function install_boost() {
         $SUDO apt-get install -y python-dev libbz2-dev
         b_ver=$1
         boost_ver=$(echo $b_ver | tr "." "_")
-        cd /tmp
-        wget http://downloads.sourceforge.net/project/boost/boost/${b_ver}/boost_${boost_ver}.tar.bz2 && \
-            tar xvjf ./boost_${boost_ver}.tar.bz2 && \
-            cd boost_${boost_ver} && \
-            ./bootstrap.sh --prefix=/usr/local && \
-            ./b2 --build-type=minimal link=shared runtime-link=shared variant=release && \
-            $SUDO ./b2 install --build-type=minimal variant=release \
-                  link=shared runtime-link=shared || \
-                die "failed to install boost"
-        cd /tmp && rm -rf boost_${boost_ver}
+        
+        install_pkg=true
+        if [ -f /usr/local/include/boost/version.hpp ]; then
+            installed_ver=$(grep -m 1 "define BOOST_LIB_VERSION" /usr/local/include/boost/version.hpp | cut -d ' ' -f 3 | tr -d '"')
+        fi
+        if [ ! -z "$installed_ver" ]; then
+            boost_ver_trim=${boost_ver%_0}
+            if [[ "$installed_ver" > "$boost_ver_trim" ]] || [[ "$installed_ver" = "$boost_ver_trim" ]]; then
+                echo "Installed boost version is sufficient: $installed_ver"
+                install_pkg=false
+            else
+                echo "Installed boost version is not sufficient: $installed_ver, proceeding with installing required bootst: $boost_ver"
+            fi
+        fi
+
+        if [ $install_pkg = true ]; then
+            cd /tmp
+            wget http://downloads.sourceforge.net/project/boost/boost/${b_ver}/boost_${boost_ver}.tar.bz2 && \
+                tar xvjf ./boost_${boost_ver}.tar.bz2 && \
+                cd boost_${boost_ver} && \
+                ./bootstrap.sh --prefix=/usr/local && \
+                ./b2 --build-type=minimal link=shared runtime-link=shared variant=release && \
+                $SUDO ./b2 install --build-type=minimal variant=release \
+                      link=shared runtime-link=shared || \
+                    die "failed to install boost"
+            cd /tmp && rm -rf boost_${boost_ver}
+        fi
     else
         boost_libs="libboost-dev libboost-graph-dev libboost-test-dev libboost-program-options-dev libboost-system-dev libboost-filesystem-dev libboost-iostreams-dev libboost-thread-dev"
         $SUDO apt-get install -y ${boost_libs}
